Rename is-path-value to isPathValue.
