+++ lib/Model.js
@@ -17,7 +17,6 @@ var arraySlice = require("./support/array-slice");
 var collectLru = require("./lru/collect");
 var pathSyntax = require("falcor-path-syntax");
 
-var isPathValue = require("./support/is-path-value");
 var isPrimitive = require("./support/is-primitive");
 var isJsonEnvelope = require("./support/is-json-envelope");
 var isJsonGraphEnvelope = require("./support/is-json-graph-envelope");
@@ -27,6 +26,7 @@ var setJsonGraphAsJsonDense = require("./set/set-json-graph-as-json-dense");
 var getSize = require("./support/getSize");
 var isObject = require("./support/isObject");
 var isFunction = require("./support/isFunction");
+var isPathValue = require("./support/isPathValue");
 var jsong = require("falcor-json-graph");
 var ID = 0;
 var validateInput = require("./support/validate-input");
