+++ lib/response/IdempotentResponse.js
@@ -11,10 +11,10 @@ var collectLru = require("./../lru/collect");
 var arrayClone = require("./../support/array-clone");
 
 var isArray = Array.isArray;
-var isPathValue = require("./../support/is-path-value");
 var isJsonEnvelope = require("./../support/is-json-envelope");
 var isJsonGraphEnvelope = require("./../support/is-json-graph-envelope");
 var isFunction = require("./../support/isFunction");
+var isPathValue = require("./../support/isPathValue");
 
 function IdempotentResponse(subscribe) {
     Observable.call(this, subscribe);
