+++ lib/support/clone.js
@@ -1,20 +1,18 @@
-var isObject = require("./../support/is-object");
 var prefix = require("./../internal/prefix");
+var hasOwn = require("./../support/hasOwn");
+var isArray = Array.isArray;
+var isObject = require("./../support/isObject");
 
 module.exports = function clone(value) {
-    var dest = value,
-        src = dest,
-        i = -1,
-        n, keys, key;
+    var dest = value;
     if (isObject(dest)) {
-        dest = {};
-        keys = Object.keys(src);
-        n = keys.length;
-        while (++i < n) {
-            key = keys[i];
-            if (key[0] !== prefix) {
-                dest[key] = src[key];
+        dest = isArray(value) ? [] : {};
+        var src = value;
+        for (var key in src) {
+            if (key[0] === prefix || !hasOwn(src, key)) {
+                continue;
             }
+            dest[key] = src[key];
         }
     }
     return dest;
