+++ test/falcor/get/get.dataSource-only.spec.js
@@ -4,7 +4,7 @@ var Rx = require('rx');
 var noOp = function() {};
 var LocalDataSource = require('../../data/LocalDataSource');
 var ErrorDataSource = require('../../data/ErrorDataSource');
-var isPathValue = require("./../../../lib/support/is-path-value");
+var isPathValue = require("./../../../lib/support/isPathValue");
 var expect = require("chai").expect;
 var sinon = require('sinon');
 var cacheGenerator = require('./../../CacheGenerator');
